Move UI code to ui/ package.
