+++ src-electron/main-process/electron-main.js
@@ -20,7 +20,7 @@ const { app } = require('electron')
 const dbApi = require('../db/db-api.js')
 const args = require('../util/args.js')
 const env = require('../util/env.js')
-const windowJs = require('./window.js')
+const windowJs = require('../ui/window.js')
 const startup = require('./startup.js')
 
 env.versionsCheck()
