+++ src-electron/main-process/startup.js
@@ -24,7 +24,7 @@ const dbEnum = require('../../src-shared/db-enum.js')
 const args = require('../util/args.js')
 const env = require('../util/env.js')
 const zclLoader = require('../zcl/zcl-loader.js')
-const windowJs = require('./window.js')
+const windowJs = require('../ui/window.js')
 const httpServer = require('../server/http-server.js')
 const ipcServer = require('../server/ipc-server.js')
 const ipcClient = require('../client/ipc-client.js')
