+++ src-electron/ui/tray.js
